+++ tests/passes/decomposer/test_yzy_decomposer.py
@@ -8,7 +8,7 @@ from opensquirrel import CNOT, CR, H, I, Rx, Ry, Rz, S, X, Y
 from opensquirrel.ir.semantics import BlochSphereRotation
 from opensquirrel.ir.single_qubit_gate import SingleQubitGate
 from opensquirrel.passes.decomposer import YZYDecomposer
-from opensquirrel.passes.decomposer.general_decomposer import check_gate_replacement
+from opensquirrel.passes.decomposer.general_decomposer import check_gate_decomposition
 
 
 @pytest.fixture
@@ -44,7 +44,7 @@ def test_yzy_decomposer(
     decomposer: YZYDecomposer, gate: SingleQubitGate, expected_result: list[SingleQubitGate]
 ) -> None:
     decomposed_gate = decomposer.decompose(gate)
-    check_gate_replacement(gate, decomposed_gate)
+    check_gate_decomposition(gate, decomposed_gate)
     assert decomposer.decompose(gate) == expected_result